--- src/main/java/sgyj/programmers/yeji/dp/Solution42895.java
@@ -1,52 +1,0 @@
-package sgyj.programmers.yeji.dp;
-
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-// N으로 표현
-public class Solution42895 {
-    public static int solution(int N, int number) {
-        List<Set<Integer>> countList = new ArrayList<>();
-
-        for(int i=0; i<9; i++)
-            countList.add(new HashSet<>());
-
-        countList.get(1).add(N); // N을 1개 쓴 값은 N 혼자이다.
-
-        for(int i=2; i<9; i++){
-            Set<Integer> countSet = countList.get(i);
-
-            for(int j=1; j<=i; j++){
-                Set<Integer> preSet = countList.get(j);
-                Set<Integer> postSet = countList.get(i - j);
-
-                for(int preNum : preSet){
-                    for(int postNum : postSet){
-                        countSet.add(preNum + postNum);
-                        countSet.add(preNum - postNum);
-                        countSet.add(preNum * postNum);
-
-                        if(preNum != 0 && postNum != 0)
-                            countSet.add(preNum / postNum);
-                    }
-                }
-            }
-
-            countSet.add(Integer.parseInt(String.valueOf(N).repeat(i)));
-        }
-
-        for(Set<Integer> sub : countList){
-            if(sub.contains(number))
-                return countList.indexOf(sub);
-        }
-
-        return -1;
-    }
-    public static void main ( String[] args ) {
-        int N = 2;
-        int number = 11;
-        System.out.println( solution(N,number) );
-    }
-}
