--- src/main/java/sgyj/programmers/yeji/hash/Solution42578.java
@@ -1,24 +1,0 @@
-package sgyj.programmers.yeji.hash;
-
-import java.util.HashMap;
-import java.util.Map;
-
-// 위장
-public class Solution42578 {
-    public int solution(String[][] clothes) {
-        int answer = 1;
-        Map<String,Integer> map = new HashMap<>();
-        for(String[] cloth : clothes){
-            map.put(cloth[1],map.getOrDefault(cloth[1],0)+1);
-        }
-        for(String key : map.keySet()){
-            answer *= (map.get(key)+1);
-        }
-
-        return answer-1;
-    }
-
-    public static void main ( String[] args ) {
-
-    }
-}
