+++ src/main/java/sgyj/programmers/yeji/section1/Solution42578.java
@@ -1,0 +1,24 @@
+package sgyj.programmers.yeji.section1;
+
+import java.util.HashMap;
+import java.util.Map;
+
+// 위장
+public class Solution42578 {
+    public int solution(String[][] clothes) {
+        int answer = 1;
+        Map<String,Integer> map = new HashMap<>();
+        for(String[] cloth : clothes){
+            map.put(cloth[1],map.getOrDefault(cloth[1],0)+1);
+        }
+        for(String key : map.keySet()){
+            answer *= (map.get(key)+1);
+        }
+
+        return answer-1;
+    }
+
+    public static void main ( String[] args ) {
+
+    }
+}
