+++ src/main/java/sgyj/programmers/yeji/section5/Solution42839.java
@@ -1,0 +1,22 @@
+package sgyj.programmers.yeji.section5;
+
+import java.util.HashSet;
+import java.util.Set;
+
+public class Solution42839 {
+
+    private static Set<Integer> primeSet = new HashSet<>();
+    private static boolean[] visited;
+    private static Character[] cm;
+    private static int answer = 0;
+    public static int solution(String numbers) {
+
+
+        return answer;
+    }
+
+    public static void main ( String[] args ) {
+        String numbers = "17";
+        System.out.println(solution(numbers));
+    }
+}
